support the --delete flag on forget and cosmetic adjustments

+++ src/drkns/context/forget.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Optional, Tuple
 
 import os
 import sys
@@ -8,25 +8,30 @@ from drkns.context.get_unit_step_path import \
     _get_path_to_unit, persistence_directory
 
 
+_to_short_extra_pattern_error_message = \
+    'Extra pattern is too short'
+
 _empty_warning_message = \
     'No files found to forget. You can use `drkns list` to see unit names'
 
 
-def forget(unit_name: Optional[str] = None):
-    clean_path = persistence_directory
-    if unit_name is not None:
-        clean_path = _get_path_to_unit(unit_name)  # type: ignore
+def forget(unit_and_pattern_input: str):
+    unit_name, extra_pattern = \
+        _get_unit_name_and_extra_pattern(unit_and_pattern_input)
+
+    clean_path = _get_path_to_unit(unit_name)
 
     if not os.path.exists(clean_path):
         sys.stderr.write(_empty_warning_message)
         return
 
     # rmtree(clean_path)  # but aws s3 sync does not remove empty directories
-
     erased_something = False
-
     for root, dirs, files in os.walk(clean_path):
         for file in files:
+            if not file.startswith(extra_pattern):
+                continue
+
             file_path = os.path.join(root, file)
             erased_something = True
             os.remove(file_path)
@@ -39,3 +44,16 @@ def forget(unit_name: Optional[str] = None):
 
     if not erased_something:
         sys.stderr.write(_empty_warning_message)
+
+
+def _get_unit_name_and_extra_pattern(
+    unit_and_pattern_input: str
+) -> Tuple[str, str]:
+    parts = unit_and_pattern_input.split('/')
+    unit_name = parts.pop(0)
+    extra_pattern = parts.pop(0) if len(parts) > 0 else ''
+    if (len(extra_pattern) > 0) and (len(extra_pattern) < 12):
+        sys.stderr.write(_to_short_extra_pattern_error_message)
+        sys.exit(1)
+
+    return unit_name, extra_pattern
